role based authentication

+++ src/main/java/com/hart/social/controller/v1/PostsController.java
@@ -1,13 +1,19 @@
 package com.hart.social.controller.v1;
 
 import com.hart.social.model.Post;
+import com.hart.social.model.User;
 import com.hart.social.repository.PostsRepository;
+import com.hart.social.repository.UsersRepository;
+import com.hart.social.security.JwtUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
 import java.util.List;
 
 @RestController
@@ -17,6 +23,12 @@ public class PostsController {
 
     @Autowired
     PostsRepository postsRepository;
+
+    @Autowired
+    UsersRepository usersRepository;
+
+    @Autowired
+    JwtUtil jwtUtil;
 
     @GetMapping(value = "/posts/{id}", produces = {"application/json"})
     public ResponseEntity getAPost(@PathVariable String id) {
@@ -40,21 +52,45 @@ public class PostsController {
     }
 
     @PostMapping(value = "/posts", produces = {"application/json"})
-    public ResponseEntity createAUser(@RequestBody Post post) {
+    public ResponseEntity createAUser(@RequestBody Post post, HttpServletRequest request) {
 
-        post.setCreatedAt(String.valueOf(System.currentTimeMillis()));
-        post.setUpdatedAt(String.valueOf(System.currentTimeMillis()));
-        postsRepository.save(post);
-        return new ResponseEntity(post, HttpStatus.ACCEPTED);
+        String token = getTokenFromRequest(request.getHeader("Authorization"));
+        User authenticatedUser = usersRepository.findUserByEmail(jwtUtil.extractUsername(token));
+
+        if (authenticatedUser.getId().equals(post.getUserId()) == true ||
+                authenticatedUser.getRole().equals("admin") == true) {
+
+            post.setCreatedAt(String.valueOf(System.currentTimeMillis()));
+            post.setUpdatedAt(String.valueOf(System.currentTimeMillis()));
+            postsRepository.save(post);
+            return new ResponseEntity(post, HttpStatus.ACCEPTED);
+
+        }
+
+        return new ResponseEntity("you are not allowed to post from another users account",
+                HttpStatus.FORBIDDEN);
+
     }
 
     @PutMapping(value = "/posts/{id}", produces = {"application/json"})
-    public ResponseEntity editAPost(@RequestBody Post post, @PathVariable String id) {
+    public ResponseEntity editAPost(@RequestBody Post post, @PathVariable String id,
+                                    HttpServletRequest request) {
 
-        post.setId(Long.valueOf(id));
-        post.setUpdatedAt(String.valueOf(System.currentTimeMillis()));
-        postsRepository.save(post);
-        return new ResponseEntity(post, HttpStatus.ACCEPTED);
+
+        String token = getTokenFromRequest(request.getHeader("Authorization"));
+        User authenticatedUser = usersRepository.findUserByEmail(jwtUtil.extractUsername(token));
+
+        if (authenticatedUser.getId().equals(post.getUserId()) == true ||
+                authenticatedUser.getRole().equals("admin") == true) {
+
+            post.setId(Long.valueOf(id));
+            post.setUpdatedAt(String.valueOf(System.currentTimeMillis()));
+            postsRepository.save(post);
+            return new ResponseEntity(post, HttpStatus.ACCEPTED);
+        }
+
+        return new ResponseEntity("you are not allowed to edit from another users account",
+                HttpStatus.FORBIDDEN);
     }
 
     @PutMapping(value = "/posts/{id}/like", produces = {"application/json"})
@@ -78,11 +114,27 @@ public class PostsController {
     }
 
     @DeleteMapping(value = "/posts/{id}", produces = {"application/json"})
-    public ResponseEntity deleteAPost(@PathVariable String id) {
+    public ResponseEntity deleteAPost(@PathVariable String id, HttpServletRequest request) {
 
-        postsRepository.deleteById(Long.valueOf(id));
-        return new ResponseEntity("Deleted",HttpStatus.ACCEPTED);
+        Post post = postsRepository.findPostById(Long.valueOf(id));
+
+        String token = getTokenFromRequest(request.getHeader("Authorization"));
+        User authenticatedUser = usersRepository.findUserByEmail(jwtUtil.extractUsername(token));
+
+        if (authenticatedUser.getId().equals(post.getUserId()) == true ||
+                authenticatedUser.getRole().equals("admin") == true) {
+
+            postsRepository.deleteById(Long.valueOf(id));
+            return new ResponseEntity("Deleted",HttpStatus.ACCEPTED);
+        }
+
+        return new ResponseEntity("you are not allowed to edit from another users account",
+                HttpStatus.FORBIDDEN);
     }
 
+    private static String getTokenFromRequest(String authorization) {
+        String token = authorization.substring(7, authorization.length());
+        return token;
+    }
 
 }
